Update namespace
The issue behind why the Client constant couldn't be loaded was because
the namespaces were structured in a way that confuses zeitwerk.

In the public-website version the structure reflected that module not
the constant.

So for example `contentful/graphql.rb` rather than
`contentful/graphql/client.rb` as it's the module that's being extended
here.

Restructuring this allows the autoloader to do its thing.