feat: add expected page url into routes
Requests to this app are controlled by CloudFront in the public-website-config repo.
The original request is sent to this app by CloudFront, so we need this app to respond to
the required path (defined by the constant `APP_PATH` in the routes.rb file).
We are effectively hard-coding the position of the app in the hierarchy of the site,
so if a content designer wants to move the app page, we will need to:
 - update the `APP_PATH` url in `routes.rb`
 - update the CloudFront config
 - add a redirect