fix: remove `dl`s from supplier detail page
An accessibility audit identified that the description lists on the supplier details page (in Other Info) were not being used as they expected.  Instead of making significant changes to the content model and the upload tool, we decided to remove the `dl`s completely and render plain markup instead.  The definition list was not really adding anything to this data in its current format.