feat: add new ECR repo for energy apps
This CDK code will eventually replace the current code in `infrastructure/cdk`.  The new infrastructure and helm deployments will contain the new name (`energy-apps`) and will be created incrementally - this PR just creates the new ECR repo.  Eventually we will switch over to the new infrastructure and delete the old infrastructure and code.